@@ -1,8 +1,9 @@
 package com.duke.dfileselector.activity;
 
 import android.Manifest;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.duke.dfileselector.util.DynamicPermissionCompat;
 
