@@ -3,12 +3,13 @@ package com.duke.fileselector;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.duke.dfileselector.activity.DefaultSelectorActivity;
 import com.duke.dfileselector.util.FileSelectorUtils;
